requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic-settings==2.10.1
redis==6.4.0
openai==1.108.1
python-dotenv==1.1.1
